+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 4.5.1
 
 ### Bug Fixes